requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow-cpu==2.16.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
gunicorn==22.0.0
